+++ jeeves/core/plugins/admin.py
@@ -15,20 +15,20 @@ class HelpPlugin(CommandPlugin):
         tokenized_msg = message.split()
 
         if tokenized_msg and tokenized_msg[0] == "list_commands":
-            self.say(channel, "%s: Here's a list of all of the available commands:" % nickname)
+            self.say(nickname, "Here's a list of all of the available commands:")
             for key in self.handler._command_plugins.keys():
-                self.say(channel, "Jeeves: %s" % key)
+                self.say(nickname, "Jeeves: %s" % key)
         elif tokenized_msg and tokenized_msg[0] in self.handler._command_plugins.keys():
             for text in self.handler._command_plugins[tokenized_msg[0]].help_text:
-                self.say(channel, text)
+                self.say(nickname, text)
         else:
             for text in self.help_text:
-                self.say(channel, text)
+                self.say(nickname, text)
 
     @property
     def help_text(self):
         help_text = [
-            "usage: Jeeves: [command] msg",
+            "usage: Jeeves: [help|-h] [list_commands] [command_name]",
             "To list available plugins `Jeeves: help list_commands`",
             "To get specific help for a plugin `Jeeves: help [command_name]`",
         ]